+++ hackerrank/DP/tests/CoinChange.cpp
@@ -6,7 +6,7 @@ namespace hackerrank
 
 TEST(CoinChange, test1)
 {
-    std::vector<long> srcArray{1, 2, 3};
+    std::vector<long long> srcArray{1, 2, 3};
     ASSERT_EQ(getWays(4, srcArray), 4);
 }
